+++ wsutil/ws_assert.c
@@ -15,7 +15,7 @@
 #include "wslog.h"
 
 
-void ws_assert_failed(const char *file, int line, const char *function,
+void ws_assert_failed(const char *file, long line, const char *function,
                     const char *domain, const char *assertion,
                     bool unreachable)
 {
